feat(python): all operators are all free functions now
Adding operators as methods of the input type doesn't lead to
a good API when we might have operators coming from different modules.
When calling an operator as a member of the input, it's not clear from
which module the operator comes from.

pillow uses a weird mix of both strategies, which we'd better avoid.

CVCUDA-276