fix: tests in Python generating intermittent CUDA errors
Errors like "cudaErrorIllegalAddress" and "[700] cuMemAlloc
UNKNOWN_CUDA_ERROR" appear in Python tests due to the usage
of user streams inside the sole test body, making some operations
run in parallel with other operations that do not use the same stream.
Since the tests are not properly unit separated, the compromise solution
is to synchronize the default stream prior to use the user stream.

CVCUDA-245