fix(python): segfault due to wrapper containers not being held by cache
When wrappers were used in a operation, their references were being
held by the cuda stream. If python script don't hold references to them
anymore, when the cuda stream is done with the wrappers, they would
get destroyed *from inside the cuda stream*.

This is a big no-no, as it cuda stream runs in a different thread,
container destruction might invoke cuda API calls, etc.

Instead, we make sure the wrappers are added to the cache. Since the
cache always holds them, they won't be destroyed inside a cuda stream.

We do some "garbage colletion" whenever we create a new wrapper in order
to avoid the number of wrappers alive exploding.

A much better solution is to have a proper memory allocator singleton,
granting us a better control of when memory is allocated and
deallocated. This will happen later.

Closes CVCUDA-344