feat: rename ndim to rank in C/C++
Python still uses ndim as it's consistent with numpy.
In C/C++, known libraries I know uses "rank" instead
(GXF, MatX, among others).

The rationale here is for the API of a language be consistent
in its own environment, to avoid surprises to users of said language.
It's more common for one to stick to one language than to be going
back and forth different languages.

CVCUDA-184