fix: make HasObjManager more robust when not using LTO (gcc-9)
Now HasObjManager doesn't depend on including the object manager
header to make it work. In some instances we were forgetting to do it,
and ToCoreObject was working with the handle as if it weren't
managed by its handle manager, leading to segfaults when we're not
using LTO, like in some files with gcc-9.

If we now create a manager and forget to include a header, the compiler
will issue an error.

Note: from now on, when an object handle manager is created, we must
specialize HandleTraits class in ICoreObject properly.

CVCUDA-506